Add Vagrant VM and Docker container
Even though we are using create-react-app to bootstrap the project, we
still want it to conform to our development environment standards of
using an Vagrant VM to run Docker containers. The changes here are a
merge of the civic apps app template with some improvements from one of
our more recent projects.