Remove redux-map-gl, use custom action and reducer
It turns out this library is out-of-date, and hasn't kept up with
changes in react-map-gl. For example, react-map-gl moved the height and
width of the map to be part of the viewport state, while redux-map-gl
was not updated with this change. This caused problems when trying to
adjust these values.

Refs #21