Implement sensor overview component
Adds a component for the sensor overview pane, which is rendered when
a sensor marker is tapped on the map.

Some adjustments are made to CSS rules for the map to fix various
issues that appear as the app layout changes. In general, it was easier
to make the map size consistent through the app and make adjustments
elsewhere.

The data for the sensor locations is now part of the state so that is
accessible to other parts of the app besides the map component.

Refs #21