Implement sensor details modal
Adds the sensor details modal to the app. It is relatively static at
this point, besides the name of the sensor, but will be wired up in #41.

The component is implemented as a React Portal, which allows us to
insert it at the top level of the app DOM instead of as a child of the
overview component so that is can be displayed correctly.

Refs #22