Add tests for sensor utility functions
In preparation for adding more sensor utility functions and wanting to
take a test-driven approach to adding those, I added a few tests for
existing utility functions.

I also modified the functions to make them more testable. This largely
involved passing the functions only data that is required by the
function, as well as data that is in a format that is directly useable
by the function.

The placeholder app tests file is removed here as well because it was
only kept around for reference purposes.

Refs #60