Poll most recent known data from live sensors
The live sensor API returns only the most recent reading taken if a time frame is not specified. We are seeing junk/no data since 12/2018 for both live sensors. Given a start date, the API sends back a list of only readings with real data since then, plus 1 additional row for the most recent sensor reading, if it is no data. Read this list in reverse order and check for no data values to grab the most recent, real, live sensor data.

Refs #59