Calculate sensor water quality ratings
Add functionality to transform the sensor data into water quality
ratings which can be used by other parts of the app to convey the health
of each waterbody.

A few tests for the water quality ratings are added for good measure.

The healthy ranges are placeholders for the moment as we await real data
from the client.

Refs #60