Adjust positioning of bottom map elements
Following up #55, now that the map height is consistent across devices,
the position of the map controls along the bottom of the map are
adjusted. The controls are positioned so that they are above the footer
on the overview map and along the bottom of the map when the sensor pane
is open.

Refs #75