Implement intro screen transition
Implements a transition from the intro screen to the home screen, as
shown in the prototype, via the following changes:

- The background fill on the intro screen was eliminated so that the map
is visible.
- A new map style was created for the intro screen that does not have
any labels. When transitioning, this style is swapped out for the style
with the labels.
- The sensor markers are only added to the map after the "Find out"
button is clicked.
- The map pitch is adjusted as the transition is happening.

Additonally, the style of the intro screen was adjusted to remove a
"hiccup" when transitioning to the home screen style due to use of
absolute positioning.

Refs #82